deepen: collapse cron sec-filing-service shallow adapter
Absorbs `lib/cron/sec-filing-service.ts` (128 LOC static class with one
method) into the single production caller, `lib/cron/handlers/discovery-handler.ts`.

## Deepening in LANGUAGE.md vocabulary

Replaces one shallow module (`CronSecFilingService.checkForNewFilings`)
with a private function inside the caller. The deleted module's own
docstring admitted the shape: 'static-class shape is retained so the
caller's existing invocation site doesn't have to change.' That is the
textbook description of a shallow adapter kept for API-preservation
rather than depth.

## Leverage callers gain

None materialise, because there was one caller. What changes is that a
maintainer chasing 'how does discovery decide a filing is new?' now finds
the loop inline in the discovery handler beside the enrichment and
bulk-job stages, instead of hopping to a 128-LOC sibling file that
wrapped a for-loop around `checkTickerForNewFilings`.

## Locality maintainers gain

The **Filing** discovery concept collapses to one place. The discovery
handler already owned the tickers→CIK enrichment (`enrichTickersWithCik`)
and the users→jobs bulk creation (`createBulkFetchJobs`); the RSS-check
loop now sits between them where it belongs. No parallel `CronSecFilingService`
class competes for the 'how do we walk tickers for new filings' concept.

## Dependency category (per process/Deepening/deepening.md)

**In-process.** The deleted module was a for-loop over
`checkTickerForNewFilings` from `lib/sec-edgar/ticker-monitoring.ts`.
The real seam to SEC EDGAR sits inside that module; the deleted class
was a hypothetical seam next to it (one adapter, one caller — per
LANGUAGE.md, that is not a seam).

## Adapter strategy

No adapter change. The inlined `discoverNewFilingsForTickers` calls
`checkTickerForNewFilings` directly — the canonical **SEC EDGAR Access**
seam named in CONTEXT.md.

## Tests deleted and tests added at the new interface

- **Deleted at the module seam:** none — the deleted module had no
  dedicated test file. Its `checkForNewFilings` interface was only
  exercised via jest.mock() in downstream tests.
- **Rewired at the discovery seam:**
  - `__tests__/cron/handlers/discovery-unprocessed.test.ts` — mock now
    targets `checkTickerForNewFilings` from
    `lib/sec-edgar/ticker-monitoring` (the underlying seam) plus
    `tickerMonitoring.findFirst` on the Prisma mock, replacing the
    per-test `CronSecFilingService.checkForNewFilings` mock. Fixture
    filings switched from the deleted `FilingWithTicker` shape to
    `RSSFilingEntry`, which is what the real SEC EDGAR seam returns.
    All 4 tests pass.
  - `__tests__/lib/cron/handlers/discovery-handler-bulk-jobs.test.ts` —
    the defensive top-level `jest.mock('sec-filing-service', ...)` block
    is removed; the test only exercised `createBulkFetchJobs` and never
    reached the mock. 3 tests pass.
  - `tests/security/cron-security.test.ts` — a stale
    `jest.mock('sec-filing-service', { discoverNewFilings, processFilingsForUser, runSecFilingMonitoring })`
    block whose methods hadn't existed for months is removed. 11 tests
    pass.

## CONTEXT.md

Updated the **SEC EDGAR Access** entry to drop the trailing
`sec-filing-service.ts` reader from the `ticker-monitoring.ts` caller
list.

## ADRs respected

- Respects ADR-0001..ADR-0006.
- No new ADR — this is a deletion of a self-admitted shallow adapter, not
  a load-bearing architectural decision.
- Same shape as PR #818 (PreferenceService slim), PR #789 (dead lib/parsers
  orbital), PR #826 (fast-poll-handler): the caller's shape didn't
  demand the intermediate class.

## Verification

- `npm test` with the CI PR-validation ignore pattern: **231 test
  suites pass, 3094 tests pass, 28 skipped, 6 todo.** The one failing
  suite (`__tests__/integration/supabase-cutover.test.ts`) fails
  identically on `main` — it asserts on `DATABASE_URL` matching a
  Supabase URL and is unrelated to this change.
- Diff: 6 files, +151 / -186 (net -35 LOC on production; a further -128
  LOC from the deleted module).
