+++ CONTEXT.md
@@ -79,7 +79,7 @@ The post-fetch gate that confirms a downloaded SEC document matches the [Filing]
 PostHog feature-flag gates that govern the [Filing Prompt] enrichment path: the top-level `why_it_matters_enrichment` kill-switch, per-provider flags (`why_it_matters_counterparty`, `why_it_matters_governance`, `why_it_matters_debt`, `why_it_matters_earnings`, `why_it_matters_capital_return`, `x_sentiment_enrichment`), and the `enable_earnings_mini_deep_dive` cohort gate that toggles the `materialitySignal` field on 10-K/10-Q schemas. Implemented as private functions inside the Summarize module (`lib/ai/summarize.ts`) — `isWhyItMattersEnabled`, `isProviderEnabled`, `isEarningsMiniDeepDiveEnabled` — composed over a single `evaluateEnrichmentFlag(flagKey, accessionNumber)` helper that dynamic-imports `getServerPostHog` so posthog-node's Node-only deps don't leak into client bundles. `accessionNumber` flows through as PostHog's `distinctId` for stable per-filing bucketing. Every gate fails-safe-off — any thrown error or missing PostHog client resolves to `false`, so an outage cannot accidentally enable enrichment. The `ENRICHMENT_FORCE_ENABLE=1` env var is a dev/QA short-circuit; production callers must never set it. Replaces a former `lib/ai/enrichment-flags.ts` (95 LOC, one production caller) whose unit tests asserted on internal flag-name constants (`_internal.TOP_LEVEL_FLAG`, `_internal.PROVIDER_FLAGS.debt_issuance`) past the interface; the integration tests in `__tests__/ai/summarize-enrichment-gate.test.ts` now exercise the gates through the Summarize seam by mocking `posthog-server`. Same shape as [Why It Matters Guard] / [Ticker Grounding Guard] / [Numeric Grounding Guard] / [Historical Context], and ADR-0002.
 
 ### SEC EDGAR Access
-Production code reaches SEC EDGAR through a small, hand-picked surface in `lib/sec-edgar/`: `client.ts` (`SECEdgarClient` — used by `lib/cron/handlers/fetch-handler.ts`), `cik-resolver.ts` (used by `app/api/user/tickers/route.ts`), `ticker-monitoring.ts` (used by the cron discovery handler and `sec-filing-service.ts`), `environment-aware-fetcher.ts` and `rss-parser.ts` (dependencies of `ticker-monitoring.ts`), `submissions-client.ts` (dependency of `client.ts`), `form-registry.{ts,js}` (kept while `services/filings/` survives), and the shared `types.ts`. There is **no barrel** — every consumer imports the specific file it needs. Replaces an orbital cluster of zero-production-caller modules that the deleted `lib/sec-edgar/index.ts` barrel had been re-exporting: `parsers.ts` (a stand-alone alternative to `filing-extractor.ts` from [Filing Extraction] that nothing chose), `filing-storage.ts` + `test-filing-storage.ts` (a `FilingStorage` class plus a manual-run REPL script — Filings are persisted through Prisma in the cron handlers directly, not through a class), and the entire `ticker-service/` subdirectory (`SECDataClient`, `TickerResolver`, its own `types.ts` — a drifting parallel to `cik-resolver.ts`, the actual production survivor). The `@/lib/sec-edgar` barrel had **zero importers** for over a year. Deletion collapses a 4-candidate hypothetical seam (4 adapters under the barrel, 0 callers) to the small file-level surface production actually exercises.
+Production code reaches SEC EDGAR through a small, hand-picked surface in `lib/sec-edgar/`: `client.ts` (`SECEdgarClient` — used by `lib/cron/handlers/fetch-handler.ts`), `cik-resolver.ts` (used by `app/api/user/tickers/route.ts`), `ticker-monitoring.ts` (used by the cron discovery handler), `environment-aware-fetcher.ts` and `rss-parser.ts` (dependencies of `ticker-monitoring.ts`), `submissions-client.ts` (dependency of `client.ts`), `form-registry.{ts,js}` (kept while `services/filings/` survives), and the shared `types.ts`. There is **no barrel** — every consumer imports the specific file it needs. Replaces an orbital cluster of zero-production-caller modules that the deleted `lib/sec-edgar/index.ts` barrel had been re-exporting: `parsers.ts` (a stand-alone alternative to `filing-extractor.ts` from [Filing Extraction] that nothing chose), `filing-storage.ts` + `test-filing-storage.ts` (a `FilingStorage` class plus a manual-run REPL script — Filings are persisted through Prisma in the cron handlers directly, not through a class), and the entire `ticker-service/` subdirectory (`SECDataClient`, `TickerResolver`, its own `types.ts` — a drifting parallel to `cik-resolver.ts`, the actual production survivor). The `@/lib/sec-edgar` barrel had **zero importers** for over a year. Deletion collapses a 4-candidate hypothetical seam (4 adapters under the barrel, 0 callers) to the small file-level surface production actually exercises.
 
 ### Monitoring
 Operational counters, gauges, and timers recorded across the cron pipeline, AI clients, and email send paths. Implemented as one deep singleton `monitoring` in `lib/monitoring/index.ts` exposing `incrementCounter` / `recordValue` / `recordGauge` / `startTimer` / `stopTimer` / `recordTiming` / `recordHealth` / `recordAiApiCall`, plus a per-metric PostHog forwarding bridge gated on the `POSTHOG_METRIC_PREFIX_ALLOWLIST` (`ai.x_sentiment_`, `ai.ticker_grounding_`, `ai.why_it_matters_`, `ai.numeric_grounding_`). Two adjacent files keep their own interfaces because they have their own production callers: `cron-monitor.ts` (`CronJobMonitor`, used by `app/api/cron/route.ts`) and `json-parsing-monitor.ts` (`jsonParsingMonitor`, used by `lib/ai/parsers/response-parser.ts`). `cron-monitor.ts` privately consumes `async-alert-queue.ts` and `performance-monitor.ts`, which are internal to the cron-alert pipeline and never reach external callers. Replaces an orbital cluster of zero-production-caller modules that the `lib/monitoring/index.ts` barrel had been re-exporting: `pipeline-error-detector.ts` (~999 LOC `PipelineErrorDetector` singleton with `ErrorPattern` / `ErrorAlert` / `PipelineMetrics` / `DetectionReport` taxonomy and a `detectPipelineErrors` / `generatePipelineHealthReport` / `resolvePipelineAlert` surface), `pipeline-health-monitoring-system.ts` (~1040 LOC `PipelineHealthMonitoringSystem extends EventEmitter` singleton with `HealthMonitoringConfig` / `HealthMetrics` / `AlertEvent` / `DashboardMetrics` taxonomy plus `recordLockOperation` / `recordParameterValidation` / `recordSecApiOperation` / `recordUserProcessing` / `getCurrentHealthMetrics` / `getDashboardMetrics` / `startPipelineHealthMonitoring` / `stopPipelineHealthMonitoring` entry points), and `sec-api-monitor.ts` (~607 LOC `MonitoredSECEdgarClient extends SECEdgarClient` adapter plus a `SecApiMonitor` class, exporting `monitoredSecClient` as the canonical instance). All three were only ever consumed by each other (pipeline-health → pipelineErrorDetector, sec-api-monitor → recordSecApiOperation) and re-exported through the barrel; every external `from '@/lib/monitoring'` import only ever destructures `{ monitoring }`. Deletion collapses a 3-candidate hypothetical seam (3 adapters under the barrel, 0 callers) to the singleton surface production actually exercises.
